Include latest migration db changes - lat/lon on site -> float

+++ api/src/main/java/au/org/aodn/nrmn/restapi/dto/site/SiteDto.java
@@ -28,11 +28,11 @@ public class SiteDto {
 
     @NotNull(message = "Longitude is required")
     @Schema(title = "Longitude")
-    private Double longitude;
+    private Float longitude;
 
     @NotNull(message = "Latitude is required")
     @Schema(title = "Latitude")
-    private Double latitude;
+    private Float latitude;
 
     @NotNull(message = "State is required")
     @Schema(title = "State")
